+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Unreleased
 
+## 2.0.0-pre3 (July 31 2020)
+
+- TRITON-2152 Added `triton accesskeys create|get|list|delete` subcommands
+
+## 2.0.0-pre2 (July 30 2020)
+
+- TRITON-2152 Added support for `account.AccessKeys`
+
 ## 2.0.0-pre1 (May 11 2020)
 
 **NOTE:** This is a precursor release of triton v2.0.0, including backwards
@@ -8,8 +16,6 @@ fields from `map[string][string]` and optionally `map[string]interface{}` to
 `map[string]interface{}` everywhere. This change suits better with the possible
 types of Metadata and Tags as defined by Triton's VMAPI: valid strings,
 booleans or numeric JSON values.
-
-- TRITON-2152 Added support for `account.AccessKeys`
 
 ## 1.8.4 (May 11 2020)